requirements.txt
streamlit
openai
plotly
gensim
nltk
openpyxl
pyLDAvis
wordcloud
python-dotenv
psycopg2-binary

